requirements.txt
Django==2.1.8
djangorestframework==3.9.3
flake8==3.6.0
mccabe==0.6.1
psycopg2==2.8.2
pycodestyle==2.4.0
pyflakes==2.0.0
pytz==2019.1
sqlparse==0.3.0
